requirements.txt
PySide6>=6.6.0
pydantic>=2.0.0
PyYAML>=6.0 